[FIX] github_connector: Reinforce deletion method
Using os.command gets hung when there's a problem deleting the folder (for example,
when no permissions), so this stucks the cron method or the manual process of
updating source code.

We avoid it using the Python utility for that through shutil.

We also avoid an exception in that case enclosing the sentence in a try/except clause.